refactor: rename internal auth package to identity
