+++ internal/identity/client_credential.go
@@ -1,4 +1,4 @@
-package auth
+package identity
 
 import (
 	"bytes"
