+++ internal/identity/identity_test.go
@@ -1,4 +1,4 @@
-package auth
+package identity
 
 import (
 	"testing"
