+++ internal/identity/managed_identity_credential.go
@@ -1,4 +1,4 @@
-package auth
+package identity
 
 import (
 	"context"
